refactor(trusty-common): split openai_compat.rs into submodules under 500-SLOC cap (#607)
Splits 523-SLOC crates/trusty-common/src/chat/openai_compat.rs into
three focused siblings under a new openai_compat/ directory:

- wire.rs      (~45 SLOC): ChatRequestWire, OpenAiToolWire, tools_wire helper
- sse_pump.rs  (~95 SLOC): ToolCallAccumulator, pump_openai_sse async fn
- providers.rs (~130 SLOC): OpenRouterProvider, OllamaProvider, auto_detect_local_provider
- mod.rs       (~160 SLOC): re-export facade + all tests (test file classification)

Zero API surface change: chat/mod.rs pub-uses remain identical.
